Time: 51 ms (69.68%), Space: 18.7 MB (67.15%) - LeetHub

+++ 0240-search-a-2d-matrix-ii/0240-search-a-2d-matrix-ii.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     bool searchMatrix(vector<vector<int>>& matrix, int target) {
-        long long n=matrix.size();
-        long long m=matrix[0].size();
+        int n=matrix.size();
+        int m=matrix[0].size();
 
         int row=0; int col=m-1;
         while(row<n && col>=0){
